match: FindNearestColor takes u8 components
The `& 0xff` at the head of FindNearestColor is the u8 -> int parameter
widening, not a source-level mask: retail's four callers push their colour
temps unwidened (GammaTable/HsvShiftTable/FlashTable byte-store then push
the whole dword; HueRampTable pushes the ftol result straight), which only
type-checks if the parameters are u8. Typing them that way removes the
caller-side masks and keeps the callee bytes.

GammaTable 94.07 -> 95.06, HsvShiftTable 94.27 -> 95.64,
FlashTable 83.82 -> 85.02, HueRampTable holds 97.45.

+++ include/DDrawMgr/ShadeTableCache.h
@@ -82,7 +82,7 @@ public:
 
     static i32 __cdecl CompareHue(const void* a, const void* b);
 
-    static i32 __cdecl FindNearestColor(PALETTEENTRY* pal, i32 r, i32 g, i32 b);
+    static i32 __cdecl FindNearestColor(PALETTEENTRY* pal, u8 r, u8 g, u8 b);
 
     i32 m_initialized;
     CShadeTableArray m_arr;
